# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SwiftEats Food Delivery Platform API
  description: |
    SwiftEats is a high-performance, scalable food delivery platform backend API.
    
    ## Key Features
    - **High-Performance Order Processing**: 500 orders/minute capacity
    - **Sub-200ms Menu Browsing**: Redis-cached restaurant and menu data
    - **Real-Time Driver Tracking**: GPS location updates for 10,000+ drivers
    - **Intelligent Driver Assignment**: Automated order-to-driver matching
    - **Live Order Tracking**: WebSocket-based real-time updates
    
    ## Performance Targets
    - Order Processing: 500 orders/minute
    - Menu Response Time: P99 < 200ms
    - Driver Location Updates: 2,000 events/second
    - Concurrent Drivers: 10,000 drivers
    
    ## Authentication
    Currently, the API operates without authentication for development/testing purposes.
    Production deployment should implement JWT-based authentication.
    
  version: 1.0.0
  contact:
    name: SwiftEats Development Team
    email: dev@swifteats.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.swifteats.com
    description: Production server
...
paths:
...
  /restaurants:
    get:
      tags: [restaurants]
      summary: Get all active restaurants
      description: Retrieve list of all active restaurants with caching for performance
      operationId: getAllActiveRestaurants
      responses:
        '200':
          description: List of active restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /restaurants/nearby:
    get:
      tags: [restaurants]
      summary: Get nearby restaurants
      description: Find restaurants within specified radius of coordinates
      operationId: getNearbyRestaurants
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 19.0760
          description: Latitude coordinate
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 72.8777
          description: Longitude coordinate
        - name: radius
          in: query
          required: false
          schema:
            type: number
            format: double
            default: 5.0
            example: 5.0
          description: Search radius in kilometers
      responses:
        '200':
          description: List of nearby restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /restaurants/{restaurantId}:
    get:
      tags: [restaurants]
      summary: Get restaurant by ID
      description: Retrieve restaurant details by ID with caching
      operationId: getRestaurantById
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '200':
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /restaurants/{restaurantId}/menu:
    get:
      tags: [restaurants]
      summary: Get restaurant menu
      description: Retrieve complete menu for a restaurant with heavy caching
      operationId: getRestaurantMenu
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '200':
          description: Restaurant menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /restaurants/{restaurantId}/menu/category/{category}:
    get:
      tags: [restaurants]
      summary: Get menu by category
      description: Retrieve menu items for a specific category
      operationId: getMenuByCategory
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - name: category
          in: path
          required: true
          schema:
            type: string
            example: "Main Course"
          description: Menu category
      responses:
        '200':
          description: Menu items in category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItemResponse'
  /restaurants/search:
    get:
      tags: [restaurants]
      summary: Search restaurants
      description: Search restaurants by name or cuisine type
      operationId: searchRestaurants
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            example: "pizza"
          description: Search term
      responses:
        '200':
          description: Matching restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantResponse'

  # Order Management APIs
...
components:
  parameters:
    RestaurantId:
      name: restaurantId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        example: 1
      description: Restaurant ID
...
  schemas:
    RestaurantResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Maharaja Restaurant"
        address:
          type: string
          example: "Mumbai Central, Mumbai"
        latitude:
          type: number
          format: double
          example: 19.0728
        longitude:
          type: number
          format: double
          example: 72.8826
        phone:
          type: string
          example: "+91-9876543210"
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BUSY, CLOSED]
          example: "ACTIVE"
        cuisineType:
          type: string
          example: "Indian"
        averageDeliveryTime:
          type: integer
          example: 30
          description: Average delivery time in minutes
        minimumOrderAmount:
          type: number
          format: double
          example: 150.0
        deliveryFee:
          type: number
          format: double
          example: 50.0
        rating:
          type: number
          format: double
          example: 4.5
        totalReviews:
          type: integer
          example: 250
    MenuItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Butter Chicken"
        description:
          type: string
          example: "Creamy tomato curry with tender chicken"
        price:
          type: number
          format: double
          example: 280.0
        category:
          type: string
          example: "Main Course"
        imageUrl:
          type: string
          example: "https://images.swifteats.com/items/1.jpg"
        available:
          type: boolean
          example: true
        preparationTime:
          type: integer
          example: 15
          description: Preparation time in minutes
        spiceLevel:
          type: string
          example: "Medium"
        isVegetarian:
          type: boolean
          example: false
        isVegan:
          type: boolean
          example: false
        calories:
          type: integer
          example: 450
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid request parameters"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T13:30:00Z"
        path:
          type: string
          example: "/api/orders"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...