# src/main/java/com/swifteats/dto/RestaurantResponseDto.java
package com.swifteats.dto;

import com.swifteats.model.Restaurant;

/**
 * DTO for Restaurant responses - optimized for fast serialization
 */
public class RestaurantResponseDto {
    
    private Long id;
    private String name;
    private String address;
    private Double latitude;
    private Double longitude;
    private String phone;
    private String status;
    private String cuisineType;
    private Integer averageDeliveryTime;
    private Double minimumOrderAmount;
    private Double deliveryFee;
    private Double rating;
    private Integer totalReviews;
    
    // Constructors
    public RestaurantResponseDto() {}
    
    public RestaurantResponseDto(Restaurant restaurant) {
        this.id = restaurant.getId();
        this.name = restaurant.getName();
        this.address = restaurant.getAddress();
        this.latitude = restaurant.getLatitude();
        this.longitude = restaurant.getLongitude();
        this.phone = restaurant.getPhone();
        this.status = restaurant.getStatus().name();
        this.cuisineType = restaurant.getCuisineType();
        this.averageDeliveryTime = restaurant.getAverageDeliveryTime();
        this.minimumOrderAmount = restaurant.getMinimumOrderAmount();
        this.deliveryFee = restaurant.getDeliveryFee();
        this.rating = restaurant.getRating();
        this.totalReviews = restaurant.getTotalReviews();
    }
    
    // Getters and Setters
    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    
    public String getAddress() { return address; }
    public void setAddress(String address) { this.address = address; }
    
    public Double getLatitude() { return latitude; }
    public void setLatitude(Double latitude) { this.latitude = latitude; }
    
    public Double getLongitude() { return longitude; }
    public void setLongitude(Double longitude) { this.longitude = longitude; }
    
    public String getPhone() { return phone; }
    public void setPhone(String phone) { this.phone = phone; }
    
    public String getStatus() { return status; }
    public void setStatus(String status) { this.status = status; }
    
    public String getCuisineType() { return cuisineType; }
    public void setCuisineType(String cuisineType) { this.cuisineType = cuisineType; }
    
    public Integer getAverageDeliveryTime() { return averageDeliveryTime; }
    public void setAverageDeliveryTime(Integer averageDeliveryTime) { this.averageDeliveryTime = averageDeliveryTime; }
    
    public Double getMinimumOrderAmount() { return minimumOrderAmount; }
    public void setMinimumOrderAmount(Double minimumOrderAmount) { this.minimumOrderAmount = minimumOrderAmount; }
    
    public Double getDeliveryFee() { return deliveryFee; }
    public void setDeliveryFee(Double deliveryFee) { this.deliveryFee = deliveryFee; }
    
    public Double getRating() { return rating; }
    public void setRating(Double rating) { this.rating = rating; }
    
    public Integer getTotalReviews() { return totalReviews; }
    public void setTotalReviews(Integer totalReviews) { this.totalReviews = totalReviews; }
}